requirements.txt
Django==3.1.2
djangorestframework==3.12.1
djangorestframework-csv==2.1.0
django-cors-headers==3.5.0
firebase-admin==4.4.0
google-cloud-firestore==1.9.0
gunicorn==20.0.4
PyJWT==1.7.1
python-dateutil==2.8.1
PyYAML==5.3.1
pyrebase==3.0.27
mysqlclient==2.0.1
uvicorn==0.12.1